Inline ContactDonationMethods, :nocov: on PlsAccount
These changes help make the coverage metric more accurate. I'm
intentionally waiting to write the specs for PlsAccount since PLS hasn't
really finalized their API. The ContactDonationMethods was just a module
to store the donation-related methods for contacts and having it in a
separate place confused the coverage report (and more generally, it's
confusing to humans and isn't a real refactoring out of functionality).