Fix contact name ordering case with Angular orderBy
A user noticed a case when a contact with a shorter last name, like
'C, J' was ordered after one with a longer last name like 'Ca, J'. It
turns out that the Postgres ordering depends on the collation. If you
used "collate 'C'" then ',' would be ordered before 'a', but with
collate 'en_US' then 'a' comes after ',' hence the strange ordering of
names in "last, first" format. Changing the collation in the SQL query
would be possible but tricky given all the Rails active record stuff we
do. A simple fix is to just add a second level of sorting in Angular.

In rare occasions it would be possible for the ordering of contacts
across pages to be inconsistent with the ordering on the page (because
the page-based ordering would use the SQL order and the contacts on the
page the Angular order), but that would be rare and less noticeable to
the user. The pages would not overlap incorrectly in anyway though.