Handle emails that differ by zero-width chars for Tnt import
The Tnt import users person.save(validate: false), which was causing a
unique key violation on (person_id, email) for EmailAddress. What
happened was that one of the email addresses in the Tnt XML export had a
zero-width space \u200B character. The check for an existing email
address in EmailAddress#add_for_person didn't strip out those chars
before checking for an existing email, hence the unique key violation.
This fixes that.