Look back 50 days from most recent donation.
From a UK user:
I know it syncs financial information once a day with dataserver.
However, does it sync everything, or only the last 30 days of
transactions? I’ve noticed quite a number of missing Gift Aid
transactions on my mpdx account. (which only come into dataserver once a
month, and often with a date older than 30 days ago). It’ll be good if
mpdx synced 50 days of transactions to avoid this.