Index whodunnit on versions
When I was checking what changes an admin fix sidekiq job did on the database I wanted to be able to query all versions created by the job and this would have been nice for it. I ended up having it filter by created_at to just get the recent ones, but even so this would be nice also if we ever need to audit what changes a user or background job did.