Don't track phone or address updates in paper trail
There has been a spike in paper trail versions recently and it
seems to be partly due to the change I made to version updates to
phone numbers. Since the phone number dup fix scripts have finished
it makes sense to turn off the versioning. Based on a quick look it
may be that there is a bug in how we do the phone number
normalization that's causing extra un-needed record changes to
phone numbers that we should maybe look into some time, but there's
no need to track extra versions for them in the mean time.

Some of the other extra version records are for addresses and it looks
like it is due to the addresses from the donor system making unnecessary
touches of address records when there are multiple addresses from the
donor system. I'll add something to JIRA to investigate that, but
similarly, no reason to save the extra version records for that right
now.