Fix intermittent CSV import bug (#777)
We had an issue where sometimes (but not always) CSV imports would not
work correctly in production. We weren't able to reproduce it locally
though which made it hard to debug.

Adam Meyer worked on it and suspected it had to do with the caching of
the import file. I investigated that a bit more and I was able to
reproduce this when I would delete the `public/uploads` folder.

I think what was happening was that when the user clicked "Run Import"
and they got served by the same server as the one where they uploaded
their import it would work because it was cached there. But if that
request got handled by the other server it wouldn't work because it
wasn't cached on that server's file system.

The solution to this is to just intentionally cache the file again as
part of the import save process.

With more work it might be possible to configure it to not need to cache
the file (since only the actual import processing needs the file
contents), but this works and the user impact of caching the file
locally should be quite small (just a small extra time loading request).