pass DataServer errors on to users
We can't predict all of the errors that DataServer might throw, for
example: You have not been setup as an MPD employee.
https://rollbar.com/Cru/MPDX/items/888/

So we can send them on to the user so if it has any useful info they can
work on that, or at least email us about it.