Reduce sidekiq logging (#893)
The infrastructure team noticed that MPDX is chewing up our Splunk
logging capacity. This really slims down Sidekiq logging a lot. We may
miss useful information but I have not recently wished for it and as we
think it would be helpful we can turn it back on to debug things.