Remove callbacks from graip generated controllers (#33)
* Remove callbacks from graip generated controllers

As so many developers are working on this project - many of whom are
unfamiliar with it, we're opting to be more explicit in our controller
actions by removing callbacks (ie: `before_action`).

While callbacks are nice to clean up controllers, they can be
ellusive.

By choosing to not use them, it means there isn't anything magical
happening behind the scenes when debugging, or working to understand
how a controller is working.

This comes at the recommendation of the GRAIP book, which specifically
emphasizes comprehension over saving lines of code (their specific
recommendation is to steer clear from functionality and gems like
[`inherited_resources`][1] and [`resource_controller`][2].

This PR adjusts the graip:controller generator to remove those generated
callbacks, instead placing them inline in the appropriate actions.

[1]: https://github.com/josevalim/inherited_resources "Inherited
Resources"

[2]: https://github.com/jamesgolick/resource_controller "Resource
Controller"

* Change load_resource to use unscoped Model#find

The initial implementation was using the scoped model to find a record
by `id` in the `load_#{resource_name}` method. Ex:

```ruby
def load_contact
  @contact ||= contact_scope.find(params[:id])
end
```

Typically this would be totally fine, as we only want to return
records that are correctly scoped. In this case, however, by scoping
our `#find` method, it would return a `404` response if the record isn't
found.

For the purpose of this API, fetching records that exist but are outside
of access for the `current_user` should instead return a `403` (which is
automatically returned after `authorize_#{resource}` fails.

```ruby
def show
  load_contact # => Fetches the contact via Contact model directly
  authorize_contact # => If no-access, returns 403

  render_contact # => Never reached if authorization fails
end
```

Note: The usage of `#{resource_name}` in the description above is due to
the generator dynamically building the methods for the resource.
Ex: `load_contacts` for `Contact`, `load_address` for `Address`, etc.