Remove app/models/appeal_donation.rb
This model doesn't have an associated table in the database, since there
is no migration to create this table.

I checked production and did not find a table in the production
database.

I also could not find any reference to AppealDonation in the code.

Since the table doesn't exist, it will cause an error when runnning
`ActiveRecord::Base.descendants`

I guess this model is not used, therefore I'm deleting it so that I can
use `ActiveRecord::Base.descendants`. :)