MPDXAPI-13 Add toontown data to seed file
I first attempted to add toontown data to the seeds by using the gem
https://github.com/rroblak/seed_dump, this approach did not work very
well because:

1. seed_dump generated 15425 lines of seeds 😳

2. The dumped seeds were not dumped in the appropriate order, to satisfy
foreign key constraints.  In order to make the dump importable, we would
have to spend hours re-ordering the seeds whenever we dump a fresh set.

I abandoned seed_dump, and attempted to seed the database using the
existing FactoryGirl factories, which seems to work.

Before this commit the test environment was using database seeds, as
well as the database_cleaner gem. They're likely a relic from Capybara
javascript tests, which don't apply to the API. The seeds have been
removed, and database_cleaner is now only used to truncate at the
beginning of the suite.