Fix specs for Docs to display JSON requests
The docs for the API are generated by using `rspec_api_documentation`
(they can be generated by running `rake docs:generate`).

When generated, however, the docs were showing the example request
bodies in the format of `application/x-www-form-urlencoded`.

This commit adjusts the `rspec_api_documentation` configuration to
convert the request body for the acceptance requests to be JSON (so that
the docs will have the correct output) as well as sets the
`Content-Type` headers in the acceptance specs to be the JSONAPI.org
media type (application/vnd.api+json).

This will in turn allow for the documentation to display true JSON
request bodies.

---

This also includes some formatting fixes