Add filter & pagination to generated controller
Now that we're using filtering and pagination for our `#index` actions -
this adjusts the generator to use said scopes on the
`#{resource_name}_scope` inside of the generated resource controller.

This also adds a `permitted_filters` method with a default value of `[]`
to ensure that an error won't be raised when loading the resources.

For more context, see commit 65e01f1.