Force UTC & iso8601 output for timestamps in Api
The goal is to have the clients convert the time to the necessary
timezone after consuming the data.

This normalizes all the API's timestamps to return as UTC iso8601.

---

Also - this monkey-patches the `Hash` class, adding the method
`#transform_values`.

This monkey-patch exists in Rails 4.2.1 already, but since we're on 4.1
I'm adding it for now.