Adjust for #update in shared_controller_examples
The shared_controller_examples allow for our controller specs to be
clean and lean. This is great! It also means that we can generate specs
that use them and save the team a ton of time.

The api does need a little massaging through. Specifically -
`correct_attributes` is used for both creating and updating resources.
This is typically fine - but I ran into issues with updating
`contact_referral` resources due to how the specs use
`correct_attributes` to create and validate against improper updates.

This adds the option to use `update_attributes` along with
`correct_attributes`. If `update_attributes` is defined as a `let`,
then those attributes will be used for the update specs. If not -
`correct_attributes` will be used instead.

Ideally - the API would be `create_attributes` and `update_attributes` -
but we're a bit under the gun to get this API out the door (and it would
require changing all the other specs.

---

This also adds the `update_attributes` declaration into the GRAIP
generator, with comments on when it should be used.