Add ErrorSerializer
This allows for the easy creating of an error response with a
resource, error hash, or error title.

If a resource or error hash is passed, the returned JSON will include a
source pointer that points to the attribute in the data hash.

The error responses follow the guidelines of the [JSONAPI spec][1].

Status is *required*

A max of one of the below is *required*

With Resource:

```ruby
serializer = ErrorSerializer.new(status: 400, resource: invalid_contact)
serializer.as_json

{
  "errors" => [
    {
      "status" => 400,
      "source" => { "pointer" => "/data/attributes/name" },
      "title" => "can't be blank"
      "detail" => "Name can't be blank"
    }
  ]
}
```

With Title:
(Detail is optional)

```ruby
arguments = {
  status: 404,
  title: 'Not Found',
  detail: 'Couldn't find User'
}

serializer = ErrorSerializer.new(arguments)
serializer.as_json

{
  "errors" => [
    {
      "status" => 404,
      "title" => "Not Found",
      "detail" => "Couldn't find User"
    }
  ]
}
```

With Hash:

```ruby
error_hash = {
  name: "can't be blank",
  email: 'must be unique'
}

serializer = ErrorSerializer.new(status: 404, hash: error_hash)
serializer.as_json

{
  "errors" => [
    {
      "status" => 404,
      "source" => { "pointer" => "/data/attributes/name" },
      "title" => "can't be blank"
      "detail" => "Name can't be blank"
    },
    {
      "status" => 404,
      "source" => { "pointer" => "/data/attributes/email" },
      "title" => "must be unique"
      "detail" => "Email must be unique"
    }
  ]
}
```

[1]: http://jsonapi.org/format/#errors "JSONAPI Error Specification"