Add ErrorSerializer usage & adjust response usage
This adds the ErrorSerializer for rendering Errors from the Api.

Also, some of our responses from controllers needed massaging.

To render out an error from a controller, check out the `render_error`
method in the `ApiController`.

Also, look at the convenience methods that wrap it, such as
`render_401`, and `render_400_with_errors`.