Fix shared examples resource_type for namespaces
This fixes the `resource_type` let statement in the shared examples for
sparse fieldsets to take into account namespaced resource types.

This `resource_type` here was being used as a key in the sparse
fieldsets fetching.

Without using a namespaced resource, this worked fine:

`User` -> `users`, which would then be used to send example field params
of:

```ruby
{
  "users": [
    "first_name",
    "last_name"
  ]
}
```

`ActiveModel::Serializers` will then use these params to ONLY return the
`first_name` and `last_name` attributes from the `UserSerializer`.

Typically this works great!

`ActiveModel::Serializers`, however, expects that top key
(ie: `"users"`) to match the "type" of the JSONAPI object that it's
generating with the `Serializer` (ie: `UserSerializer`).

Before, `Contact::Analytics` would come out as `"contact/analytics"`,
which isn't a valid `type` expected by `ActiveModel::Serializers`.

The `Contact::AnalyticsSerializer` would be instead expecting to receive
`"contact_analytics"` as the top key in the field params.

This adjusts the generation of the `resource_type` to account for the
namespace in `Contact::Analytics`, and instead return
`"contact_analytics"` so that the field params will be correctly
received and the correct fields returned.