RecordsWithSpecialDayBetweenDatesFinder service
This adds a service class to that can be used to find records that have
a "special day", ie: birthday or anniversary that will take place
between two dates.

Example:

```ruby
today = Date.current

person = create(:person, {
  birthday_day: today.day,
  birthday_month: today.month,
  birthday_year: today.year + 10.years
})

RecordsWithSpecialDayBetweenDatesFinder.find(
  scope: Person,
  attribute: :birthday,
  start_date: today - 1.day,
  end_date: today + 1.day
)

=> "#<Person::ActiveRecord_Relation:0x007fef69eced60 @records=[person]>"
```

The return object is an ActiveRecord::Relation, and as such - the
`Finder` class can be used within a scope, and its return value can be
chained with other scopes / queries.