Adjust ErrorSerializer to check resource errors
In the previous implementation, the `ErrorSerializer` would check that a
resource was passed in, and that it was invalid, before returning its
errors in the json.

This worked in theory, but when used with saving resources within a
context, ie: `@user.save(context: update_from_controller)` - the save
would fail for the right reasons, but the record itself would return
`true` when checked again if valid (as `ErrorSerializer` had no concept
of checking validity within a context).

Instead - we'll now check that the resource has been provided and that
it has errors, ie: `resource.errors.present?`.