MPDX-1928 Improvements to reports/balances (#194)
This story was created to fix/improve the balances report. Later it was
found that the data returned in the balances report is already available
on the endpoint GET /api/v2/account_lists/:account_list_id/designation_accounts
This makes the balances report redundant, so this commit removes it entirely.