Hotfix: Update user tracked info more often (#703)
We need user's tracked info to be up-to-date so that we can scope them
for the account list data import. Before this update the tracked info
would only update when a new jwt is created. But it might be up to 30
days until a user receives a new jwt token, because that is the
expiration time. This change fixes the issue by updating the user
tracked info about once per day (every 12 hours).