MPDX-4142-2
This is the second commit for MPDX-4142, this commit is concerned with
moving the donation import logic in appeals_import.rb into
gifts_import.rb, and making it work for both Tnt 3.1 and 3.2.

Awhile ago we updated the gifts_import.rb to assign the donations to
appeals, at that time we didn't know that appeals_import.rb was also
doing this. The old implementation in appeals_import.rb is out of date and
only works if the donations were already imported from the organization
before the Tnt import is run. This commit merges the older logic in
appeals_import.rb into gifts_import.rb.