bump pg and sidekiq gem versions (#1154)
In an effort to make sidekiq more stable (https://jira.cru.org/browse/MPDX-5042), I'm trying to update to the latest sub version of pg and sidekiq(pro) as we are on. PG has released version 1 and Sidekiq has released version 4, but we can save those updates for another day.