Update AppointmentResultsPeriod  to render contact changes [MPDX-4941] (#1151)
This PR Resolves MPDX-4941

The goal is to enable clients to render report like: "Contact A increased from 10 USD to EUR 50"

One interesting thing about this is that it always converts the difference as though they changed today. We could go back and user the middle of the period or the average date on the PledgeStatusLogs, but I started with this. I have asked for a business decision from Jimmy in JIRA.

MonthlyIncreaseContact was the best name I could come up with to store the meta data on these changes. I'm up for suggestions if anyone can think of a more appropriate name.

## This pull request is a

- Improvement