More import improvements (#1167)
as a continuation of #1160, this continues to dig into why some import jobs spin for 20+ hours before eventually getting killed by sidekiq restarting.

The biggest identified issue is that when a task is imported from tnt with a newsletter type, those tasks are initially being saved with no contacts and then having the activity_contact relationship added immediately after. The problem with that is that our Tasks code will treat new newsletter tasks with no contacts as though they should be duplicated to every contact: https://github.com/CruGlobal/mpdx_api/blob/bd99b39/app/models/task.rb#L443-L446

I also reduced the chances of the task counter being re-calculated when importing tasks and removed the uniqueness check of our auditing (checking for this uniqueness adds a small amount of overhead to every database transaction and we don't really care a lot if every now and then an audit is double logged).