add NewsletterTaskCleanupWorker (#1172)
for https://secure.helpscout.net/conversation/586845425/219989?folderId=354370

After #1167 we need to clean up a lot of duplicated tasks (millions in the case of this user). This would take a long time running on a dev console so I need a worker that can keep running this without someone being connected.